+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Changelog
 
-## [unreleased] - 2023-01-xx
+## [v0.5.0-alpha.27] - 2023-01-18
 
-- Fixed a bug where the list of last stati was no updated on status change
+- Fixed a bug where the list of last stati wasn't updated on status change
 - Windows only: Fixed the annoying bug that created a modal with "can't open bytedance url".
 
 ## [v0.5.0-alpha.26] - 2023-01-17